+++ parser.h
@@ -3,8 +3,8 @@
 
 #include "mpc/mpc.h"
 
-mpc_parser_t* Integer;
-mpc_parser_t* Float;
+mpc_parser_t* Long;
+mpc_parser_t* Double;
 mpc_parser_t* Operator;
 mpc_parser_t* Expr;
 mpc_parser_t* Lisp;
